created registration command and bug fixes
I updated the university command to fix a bug that was there so it now actually runs. Created the registration command that displays the beginning of registration date based on year. Updated the messageCreate to allow for args to be passed to the methods. Further updates needed to get registration working with slash commands, but for now uses !